requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
et_xmlfile==2.0.0
fastapi==0.134.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-secret-manager==2.26.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpc-google-iam-v1==0.14.3
grpcio==1.78.1
grpcio-status==1.78.1
h11==0.16.0
h2==4.3.0
holidays==0.91
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
joblib==1.5.3
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
playwright==1.58.0
postgrest==2.28.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
realtime==2.28.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
xgboost==3.2.0
yarl==1.22.0
zstandard==0.25.0
